Conflicts:
	Fproject/src/com/toleisure/mybatis/mapper/MainDAO.xml

+++ Fproject/WebContent/WEB-INF/views/eventDetail.jsp
@@ -40,7 +40,14 @@ function edit_paging(boardNum)
 
 function del_event(boardNum)
 {	
-	location.href="eventdelete.action?boardNum="+boardNum;
+	if (confirm("삭제하시겠습니까?")==true)
+	{
+		location.href="eventdelete.action?boardNum="+boardNum;	
+	} else
+	{
+		return;
+	}
+	
 }
 
 function fn_paging(curPage)
